+++ packages/js/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @postrun/js
 
+## 2.2.1
+
 ## 2.2.0
 
 ### Minor Changes